+++ FinalProject/src/main/webapp/WEB-INF/jsp/updateGuide.jsp
@@ -72,7 +72,7 @@
                         <td class="formLabel"><label class="GIFLabel">性別： </label></td>
                         <td class="formContent">
                           <div class="btn-group" role="group" aria-label="Basic radio toggle button group"
-                            style="float:right; margin-right: 50px;">
+                            style="float:right; margin-right: 70px;">
 
                             <form:radiobutton path="guideGender" value="男性" class="btn-check" name="btnradio"
                               id="btnradio1" autocomplete="off" />
